Update transport mapping for DVB-C
Per Karl Dietz, typically a Schedules Direct
DVB-C transport is provided as a XMLTV DTV
transport (with network/transport/service id).

Update the mapping accordingly.

This is a minor first step in the proper mapping
and creation of the get-lineup output which may
allow (with many limitations) of semi-automatic
creation of channels in some applications (in
particular MythTV).