Update DVB-C/T/S get-lineup output
Try to improve the output of get-lineup for DVB-C/T/S
transports to potentially allow auto-mapping for
channels.  This is a work in progress.  Thanks to
Karl Dietz for the assistance in better understanding
the details of such transports (but my understanding
is likely still incomplete).

Note that the source data for transport type from
Schedules Direct is apparently incorrect for some
lineups, so testing is based on manual database
manipulations.