hack for tv_find_grabbers source parsing of description
tv_find_grabbers will attempt to parse the source of a
grabber to determine the description in preference to
using the --description option.  While that approach
seems prone to errors (as it misparsed a variable
substitution in this grabbers case), it is what it is,
so just place the description inline to make it work.