Add supplemental (SHow) data to program (EPisode) data
Extensive update to provide "supplemental" data to
programs.  Schedules Direct EPisode (program) data
may contain only the details for the particular EPisode,
with more generic information only being available
in the SHow record.  There is no explicit reference to
the SHow record, but it can be derived.  Because there
is no reference or hash for the SHow record, we have
to manage them slightly differently, and use an aging
algorithm to randomly refresh any such supplemental
SHows so that the supplemental information is not
excessively stale.  As part of merging the supplemental
data to programs, some improvements in cast/crew
ordering and roles was performed, along with better
episode and series numbering now available.