Add support for downloading artwork.
Although programmes can have an episodeImage, this is often
missing. So we now fetch image information for every programme
we are downloading and use that to generate the programme icon.

Since many programmes and show information are already cached,
the database will need to be cleared if you want the artwork
immediately for all programmes, otherwise it will be downloaded
the next time the show is updated.