add support for Gracenote rating body advisories
Gracenote (the upstream for Schedules Direct) is
adding content advisories to the contentRating
body.  Schedules Direct has (as a value add) a
contentAdvisory set of advisories.  Support the
collection from the Gracenote ratings.  From all
appearences (reviewing the data provided) the
Gracenote roll out is still a work in progress,
so future changes in this area is likely.