remove location from lineup displays
With API-NEXT the lineup location
is not consistently available, so
just remove it from various displays.
It will still be displayed on the
guided selection choices (if available).

For this change we will not do a schema
update (to delete the location from the
lineup) in order to provide a bit of
backward compatibility.