always download current account lineups
As a lineup request is low overhead, avoid trying
to optimize it away, and also add the lineup
cleanup routine into the download process directly.

In API-NEXT all the lineup information will be in
the status response, which is already required to
be available, so it will have next to zero overhead.

Also some minor message consistency.